pyproject.toml
[project]
name = "lab-tracker"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.22.0",
    "ipykernel>=7.0.1",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.7",
    "notebook>=7.4.7",
    "pandas>=2.3.3",
    "pillow>=10.2.0",
    "pydantic-settings>=2.11.0",
    "pymupdf>=1.26.5",
    "surya-ocr>=0.17.0",
    "yandex-cloud-ml-sdk>=0.16.0",
]
